Changes package name + uses singleton for SoundManager
The convention is to use br.cefetmg.games.PACKAGE_NAME, so this commits
fixes br.cefetmg.sound to br.cefetmg.games.sound.
Also, the SoundManeger (now SoundManager - there was a typo) acts like
a singleton class instead of like a static one, which makes more sense
as there is some initialization code as soon as it is created (setting
audioEnabled to true and initializing the audios hash set).